requirements.txt
pywebview
requests
jinja2
google-genai
# optional for pdf conversion
# pdf2image 
